docker/restapi/rest_api/requirements.txt
mysql-connector-python==8.0.27
Flask==2.0.2